Centralize Scene08_MapScreenTick prototype into include/game.h
True return interface is s32 (no exit mask in binary; 2 callers declared
u32, 1 u8); def was u8 — changed to s32, byte-neutral. 3 callers
de-externed, proto inserted into game.h.

+++ src/engine/sub_0800fd50.c
@@ -27,7 +27,7 @@ extern void EntityRec_ClearKeyFlags(u8 *rec, u16 mask);
  *     compiled size is 0xC0 (192 B), not 0xA4 (164 B). The linker.ld
  *     slot is sized accordingly. */
 
-u8 Scene08_MapScreenTick(void)
+s32 Scene08_MapScreenTick(void)
 {
     u8 *state;
     u8 ret;
